requirements.txt
asgiref==3.8.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==4.2.15
django-allauth==64.1.0
django-paypal==2.1
djangorestframework==3.15.2
Flask==3.0.3
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
oauthlib==3.2.2
paypalrestsdk==1.13.3
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
South==1.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.20.0
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
